[NAYB-46] refactor : 카테고리 엔티티 패키지 구조 변경
- category/domain/MainCategory -> category/MainCategory
- category/domain/SubCategory -> category/SubCategory
