+++ src/main/java/com/prgrms/nabmart/domain/category/MainCategory.java
@@ -1,14 +1,11 @@
-package com.prgrms.nabmart.domain.category.domain;
+package com.prgrms.nabmart.domain.category;
 
 import com.prgrms.nabmart.domain.item.domain.Item;
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
-import jakarta.persistence.FetchType;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
-import jakarta.persistence.JoinColumn;
-import jakarta.persistence.ManyToOne;
 import jakarta.persistence.OneToMany;
 import java.util.ArrayList;
 import java.util.List;
@@ -19,24 +16,22 @@ import lombok.NoArgsConstructor;
 @Getter
 @Entity
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
-public class SubCategory {
+public class MainCategory {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long subCategoryId;
-
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "main_category_id")
-    private MainCategory mainCategory;
+    private Long mainCategoryId;
 
     @Column(nullable = false)
     private String name;
 
-    @OneToMany(mappedBy = "subCategory")
+    @OneToMany(mappedBy = "mainCategory")
+    private List<SubCategory> subCategories = new ArrayList<>();
+
+    @OneToMany(mappedBy = "mainCategory")
     private List<Item> items = new ArrayList<>();
 
-    public SubCategory(MainCategory mainCategory, String name) {
-        this.mainCategory = mainCategory;
+    public MainCategory(String name) {
         this.name = name;
     }
 }
