+++ src/main/java/com/prgrms/nabmart/domain/category/SubCategory.java
@@ -1,11 +1,14 @@
-package com.prgrms.nabmart.domain.category.domain;
+package com.prgrms.nabmart.domain.category;
 
 import com.prgrms.nabmart.domain.item.domain.Item;
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
+import jakarta.persistence.FetchType;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
+import jakarta.persistence.JoinColumn;
+import jakarta.persistence.ManyToOne;
 import jakarta.persistence.OneToMany;
 import java.util.ArrayList;
 import java.util.List;
@@ -16,22 +19,24 @@ import lombok.NoArgsConstructor;
 @Getter
 @Entity
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
-public class MainCategory {
+public class SubCategory {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long mainCategoryId;
+    private Long subCategoryId;
+
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "main_category_id")
+    private MainCategory mainCategory;
 
     @Column(nullable = false)
     private String name;
 
-    @OneToMany(mappedBy = "mainCategory")
-    private List<SubCategory> subCategories = new ArrayList<>();
-
-    @OneToMany(mappedBy = "mainCategory")
+    @OneToMany(mappedBy = "subCategory")
     private List<Item> items = new ArrayList<>();
 
-    public MainCategory(String name) {
+    public SubCategory(MainCategory mainCategory, String name) {
+        this.mainCategory = mainCategory;
         this.name = name;
     }
 }
